Renamed package errtestutil->errtesting
