+++ pkg/wharfyml/stage_test.go
@@ -3,7 +3,7 @@ package wharfyml
 import (
 	"testing"
 
-	"github.com/iver-wharf/wharf-cmd/internal/errtestutil"
+	"github.com/iver-wharf/wharf-cmd/internal/errtesting"
 	"github.com/iver-wharf/wharf-cmd/internal/yamltesting"
 	"github.com/iver-wharf/wharf-cmd/pkg/wharfyml/visit"
 	"github.com/stretchr/testify/assert"
@@ -12,13 +12,13 @@ import (
 func TestVisitStage_ErrIfNotMap(t *testing.T) {
 	key, node := yamltesting.NewKeyedNode(t, `myStage: 123`)
 	_, errs := visitStageNode(key, node, Args{}, nil)
-	errtestutil.RequireContainsErr(t, errs, visit.ErrInvalidFieldType)
+	errtesting.RequireContainsErr(t, errs, visit.ErrInvalidFieldType)
 }
 
 func TestVisitStage_ErrIfEmptyMap(t *testing.T) {
 	key, node := yamltesting.NewKeyedNode(t, `myStage: {}`)
 	_, errs := visitStageNode(key, node, Args{}, nil)
-	errtestutil.RequireContainsErr(t, errs, ErrStageEmpty)
+	errtesting.RequireContainsErr(t, errs, ErrStageEmpty)
 }
 
 func TestVisitStage_Name(t *testing.T) {
