+++ pkg/wharfyml/steptype_test.go
@@ -3,7 +3,7 @@ package wharfyml
 import (
 	"testing"
 
-	"github.com/iver-wharf/wharf-cmd/internal/errtestutil"
+	"github.com/iver-wharf/wharf-cmd/internal/errtesting"
 	"github.com/iver-wharf/wharf-cmd/internal/yamltesting"
 	"github.com/iver-wharf/wharf-cmd/pkg/wharfyml/visit"
 )
@@ -11,5 +11,5 @@ import (
 func TestVisitStepType_ErrIfNotMap(t *testing.T) {
 	key, node := yamltesting.NewKeyedNode(t, `container: 123`)
 	_, _, errs := visitStepTypeNode("", key, node, Args{}, testVarSource)
-	errtestutil.RequireContainsErr(t, errs, visit.ErrInvalidFieldType)
+	errtesting.RequireContainsErr(t, errs, visit.ErrInvalidFieldType)
 }
